requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astroid==2.11.3
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
dill==0.3.4
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
pep8==1.7.1
platformdirs==2.5.2
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298595717/work
pycodestyle==2.8.0
pylint==2.13.7
pyparsing==3.0.8
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
swagger-spec-validator==2.7.4
toml==0.10.2
tomli==2.0.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
